fix: device registration when user logins to the platform - EXO-64015
